# docs/api/openapi/bundled/control-api.yaml
openapi: 3.1.0
info:
  title: Dataspace Control Plane API
  description: Single backend entrypoint for human operators and automation.
  version: 0.1.0
  license:
    name: Proprietary - Internal Use Only
    identifier: LicenseRef-Proprietary-Internal-Use-Only
paths:
...
  /api/v1/streams/tickets:
    post:
      tags:
      - streams
      summary: Issue Stream Ticket
      operationId: issue_stream_ticket
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamTicketRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamTicketResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StreamTicketRequest:
      properties:
        workflow_id:
          type: string
          title: Workflow Id
      type: object
      required:
      - workflow_id
      title: StreamTicketRequest
    StreamTicketResponse:
      properties:
        ticket:
          type: string
          title: Ticket
        expires_in_seconds:
          type: integer
          title: Expires In Seconds
      type: object
      required:
      - ticket
      - expires_in_seconds
      title: StreamTicketResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: /
  description: Environment-relative deployment base URL.